Refactor the BoardingNavigationController to a BoardingPassPresenter so BoardingPass can be used with other UINavigationController subclasses
